requirements.txt
tensorflow==2.6.0
numpy==1.19.5
matplotlib==3.3.4
Pillow==8.2.0
